Extract composite container into a single file
